+++ Mandelbrot/MainFormEventArgs.cs
@@ -4,14 +4,14 @@ namespace Mandelbrot;
 
 public class MainFormEventArgs : EventArgs
 {
-	private decimal _scale;
+	private double _scale;
 	private int _quality;
 	private Vector2Int _frameSize;
 	private Vector2 _worldCenter;
 	private PictureBox _viewport;
 	private Color _color;
 
-	public MainFormEventArgs(decimal scale, int quality, Vector2Int frameSize, Vector2 worldCenter, PictureBox viewport, Color color)
+	public MainFormEventArgs(double scale, int quality, Vector2Int frameSize, Vector2 worldCenter, PictureBox viewport, Color color)
 	{
 		_scale = scale;
 		_quality = quality;
@@ -21,7 +21,7 @@ public class MainFormEventArgs : EventArgs
 		_color = color;
 	}
 
-	public decimal Scale => _scale;
+	public double Scale => _scale;
 	public int Quality => _quality;
 	public Vector2Int FrameSize => _frameSize;
 	public Vector2 WorldCenter => _worldCenter;
