+++ Mathematics/Vector2.cs
@@ -4,21 +4,21 @@ namespace Mathematics;
 
 public readonly struct Vector2
 {
-	private readonly decimal _x, _y;
+	private readonly double _x, _y;
 
 	public Vector2()
 	{
-		_x = 0.0M;
-		_y = 0.0M;
+		_x = 0.0D;
+		_y = 0.0D;
 	}
-	public Vector2(decimal x, decimal y)
+	public Vector2(double x, double y)
 	{
 		_x = x;
 		_y = y;
 	}
 
-	public decimal X => _x;
-	public decimal Y => _y;
+	public double X => _x;
+	public double Y => _y;
 
 	public static Vector2 Zero => new Vector2(0, 0);
